addOffer and showOffer screens.
